third_party/WebKit/LayoutTests/fast/html/imports/rel-style-to-import.html

<!DOCTYPE html>
<html>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/import-helpers.js"></script>
<head>
<link id="target", rel="stylesheet" href="resources/hello.css">
</head>
<body>
<script>
waitAndTest([
    {
        description: "Updating attribute value shouldn't result import loading.",
        setup: function()
        {
            debug("Before update:");
            shouldBeNonNull("target.sheet");
            shouldBeNull("target.import");
            target.setAttribute("href", "resources/hello.html");
            target.setAttribute("rel", "import");
            window.placeholder = createPlaceholder();
        },
        ready: function() { return window.placeholder.import && target.sheet; },
        test: function()
        {
            debug("After update:");
            shouldBeNonNull("target.sheet");
            shouldBeNull("target.import");
        }
    }
]);
</script>
</body>
</html>
